ENH: enabling color image support
Number of components now set from image message.
Display node color table now set depending on number of components.

I'm not sure if there's a clean way of downcasting an igtl::MessageBase into a igtl::ImageMessage so for now I operate on the inner pointer. If anything fails it reverts to the original greyscale image.