BUG: Socket connection remains open after deleting connector node
The logic class kept an observer on the connector node until the logic class was deleted (when Slicer exited), therefore socket connections were not closed until Slicer exited.

Fixed by observing removing the observer when the connector node is deleted. Now the socket connection is closed if the connector node is deleted or the scene is closed.