BUG: Fix hang when sending a blocking command in response to a connection
Sending a blocking command will attempt to call PeriodicProcess again to force the command to be sent. This can cause a deadlock if it is sent in response to an event due to the lock on EventQueueMutex.
Fixed by converting EventQueueMutex to a recursive_mutex since multiple same-thread access to EventQueue seems safe.